v2.1.5.9
spigot-api 대신 bukkit-api를 사용하도록 변경
베타 능력: 유령
스토커 스킬 사용 중 모든 플레이어의 텔레포트나 게임모드 변경이 막히던 문제 해결

+++ modules/Plugin/src/daybreak/abilitywar/ability/list/BombArrow.java
@@ -59,7 +59,7 @@ public class BombArrow extends AbilityBase {
 		}
 	}.setPeriod(TimeUnit.TICKS, WRECK.isEnabled(getGame()) ? 60 : 140);
 
-	private final int size = SizeConfig.getValue();
+	private final float size = SizeConfig.getValue();
 
 	private final ActionbarChannel actionbarChannel = newActionbarChannel();
